requirements.txt
#####
Python==3.11.3
pip==25.2
Django==5.2.6
djangorestframework==3.16.1



